+++ Source/Forged.MapServer/Chat/Commands/WPCommands.cs
@@ -66,10 +66,10 @@ internal class WPCommands
         stmt = DB.World.GetPreparedStatement(WorldStatements.INS_WAYPOINT_DATA);
         stmt.AddValue(0, pathId);
         stmt.AddValue(1, point + 1);
-        stmt.AddValue((int)2, (float)player.Location.X);
-        stmt.AddValue((int)3, (float)player.Location.Y);
-        stmt.AddValue((int)4, (float)player.Location.Z);
-        stmt.AddValue((int)5, (float)player.Location.Orientation);
+        stmt.AddValue(2, player.Location.X);
+        stmt.AddValue(3, player.Location.Y);
+        stmt.AddValue(4, player.Location.Z);
+        stmt.AddValue(5, player.Location.Orientation);
 
         DB.World.Execute(stmt);
